Fjern fullføring-knapp for kjørelistebehandling, bruk kun send til beslutter
Beholder validering av ulagrede endringer før sending.
Fjerner useFullførKjøreliste som nå er ubrukt.
