+++ src/frontend/Sider/Behandling/Kjøreliste/FullførKjørelisteFane.tsx
@@ -1,14 +1,14 @@
-import React, { FC } from 'react';
+import React, { FC, useState } from 'react';
 
 import { Button, VStack } from '@navikt/ds-react';
 
 import styles from './FullførKjørelisteFane.module.css';
 import { KjørelisteBrevmeny } from './KjørelisteBrevmeny';
-import { useFullførKjøreliste } from './useFullførKjøreliste';
 import { useKjørelisteBrev } from './useKjørelisteBrev';
 import { useBehandling } from '../../../context/BehandlingContext';
 import { useSteg } from '../../../context/StegContext';
 import { Feilmelding } from '../../../komponenter/Feil/Feilmelding';
+import { Feil, lagFeilmelding } from '../../../komponenter/Feil/feilmeldingUtils';
 import { PdfVisning } from '../../../komponenter/PdfVisning';
 import { Steg } from '../../../typer/behandling/steg';
 import { useSendTilBeslutter } from '../Brev/useSendTilBeslutter';
@@ -19,10 +19,28 @@ export const FullførKjørelisteFane: FC = () => {
     const { erStegRedigerbart } = useSteg();
 
     const { brevPdf, settBrevPdf, lagretBegrunnelse } = useKjørelisteBrev();
-    const { fullførKjøreliste, laster, feilmelding, settHarUlagredeEndringer } =
-        useFullførKjøreliste();
     const { sendTilBeslutter, visVedtakFerdigstiltModal, lukkVedtakFerdigstiltModal } =
         useSendTilBeslutter();
+
+    const [harUlagredeEndringer, settHarUlagredeEndringerIntern] = useState<boolean>(false);
+    const [feilmelding, settFeilmelding] = useState<Feil>();
+
+    const settHarUlagredeEndringer = (harEndringer: boolean) => {
+        if (!harEndringer) settFeilmelding(undefined);
+        settHarUlagredeEndringerIntern(harEndringer);
+    };
+
+    const handleSendTilBeslutter = () => {
+        if (harUlagredeEndringer) {
+            settFeilmelding(
+                lagFeilmelding(
+                    'Du har endret begrunnelsen uten å lagre. Trykk "Lagre begrunnelse" eller fjern teksten før du sender til beslutter.'
+                )
+            );
+            return;
+        }
+        sendTilBeslutter();
+    };
 
     const erSendTilBeslutterSteg = behandling.steg === Steg.SEND_TIL_BESLUTTER;
     const erRedigerbarISteg =
@@ -45,24 +63,13 @@ export const FullførKjørelisteFane: FC = () => {
 
                         {erRedigerbarISteg && (
                             <div>
-                                {erSendTilBeslutterSteg ? (
-                                    <Button
-                                        variant="primary"
-                                        onClick={() => sendTilBeslutter()}
-                                        size={'small'}
-                                    >
-                                        Send til beslutter
-                                    </Button>
-                                ) : (
-                                    <Button
-                                        variant="primary"
-                                        loading={laster}
-                                        onClick={fullførKjøreliste}
-                                        size={'small'}
-                                    >
-                                        Fullfør kjørelistebehandling
-                                    </Button>
-                                )}
+                                <Button
+                                    variant="primary"
+                                    onClick={handleSendTilBeslutter}
+                                    size={'small'}
+                                >
+                                    Send til beslutter
+                                </Button>
                             </div>
                         )}
                     </VStack>
